+++ integreat_cms/cms/migrations/0039_poi_category.py
@@ -26,7 +26,7 @@ def update_roles(apps, schema_editor):
 
     # Clear and update permissions according to new constants
     for role_name in dict(roles.CHOICES):
-        group = Group.objects.filter(name=role_name).first()
+        group, _ = Group.objects.get_or_create(name=role_name)
         # Clear permissions
         group.permissions.clear()
         # Set permissions